Move web.html route to routes.rs
